+++ lib/src/json/zones/box.g.dart
@@ -14,6 +14,10 @@ Box _$BoxFromJson(Map<String, dynamic> json) => Box(
       terrainId: json['terrainId'] as String,
       enclosed: json['enclosed'] as bool? ?? false,
       reverbId: json['reverbId'] as String?,
+      enterMessage: json['enterMessage'] == null
+          ? null
+          : CustomMessage.fromJson(
+              json['enterMessage'] as Map<String, dynamic>),
     );
 
 Map<String, dynamic> _$BoxToJson(Box instance) => <String, dynamic>{
@@ -24,4 +28,5 @@ Map<String, dynamic> _$BoxToJson(Box instance) => <String, dynamic>{
       'terrainId': instance.terrainId,
       'enclosed': instance.enclosed,
       'reverbId': instance.reverbId,
+      'enterMessage': instance.enterMessage,
     };